after deleting task, reset the list

+++ src/components/TaskOverview.vue
@@ -5,6 +5,12 @@
     </q-dialog>
   </div>-->
   <div class="row justify-start">
+    <!-- <q-chip
+      v-model:selected="desert.Icecream"
+      color="primary"
+      text-color="white"
+      icon="cake"
+    >Ice cream</q-chip>-->
     <q-chip
       v-for="(item, index) in listData"
       :key="item.title"
@@ -40,7 +46,7 @@
   </div>
 </template>
 <script>
-import { watch, computed, ref } from "vue";
+import { reactive, watch, computed, ref } from "vue";
 import { useStore } from "vuex";
 import scheduleCard from "@/components/ScheduleCard.vue"; // @ is an alias to /src
 
@@ -52,18 +58,28 @@ export default {
   setup(props) {
     const store = useStore();
 
+    const setInTimeTable = item => {
+      alert(item);
+    };
+
     const startDrag = (event, item) => {
       event.dataTransfer.setData("itemTitle", item.title);
       event.dataTransfer.setData("itemDesc", item.desc);
     };
-
-    const deleteItem = item => {
-      store.dispatch("Task/deleteTask", { id: item._id });
+    const deleteItem = async item => {
+      if (confirm(`Do you really want to delete [${item.title}]?`)) {
+        await store.dispatch("Task/deleteTask", { id: item._id });
+        await store
+          .dispatch("Task/getTaskList")
+          .then(result => console.log("성공", result))
+          .catch(error => console.error("실패", error));
+      }
     };
 
     return {
       dialog: ref(false),
       filter: ref(""),
+      setInTimeTable,
       startDrag,
       deleteItem,
       listData: computed(() => {
